+++ src/components/profile/options.tsx
@@ -1,6 +1,8 @@
 import { View, Text, Pressable } from 'react-native';
-
-const Options = () => {
+interface OptionsTypes {
+  logoutHandler: () => Promise<void>;
+}
+const Options = ({ logoutHandler }: OptionsTypes) => {
   return (
     <View className="flex flex-col items-center gap-2 text-white">
       <Pressable className="w-3/4 rounded-xl bg-tethr-purple/80 py-1">
@@ -11,7 +13,7 @@ const Options = () => {
         {' '}
         <Text> Delete Account</Text>{' '}
       </Pressable>
-      <Pressable className="my-5 w-3/4 rounded-xl bg-tethr-purple/80 py-1">
+      <Pressable className="my-5 w-3/4 rounded-xl bg-tethr-purple/80 py-1" onPress={logoutHandler}>
         <Text> Privacy Policy</Text>
       </Pressable>
     </View>
